feat(backend/sgx): factor out enarxcalls
